youtube-cli/requirements.txt
yt-dlp
google-api-python-client
youtube-transcript-api
python-dotenv
typing-extensions
